pull rolls out of VisualView
